requirements.txt
discord.py==1.5.1
discord-py-slash-command==1.0.7
python-dotenv==0.15.0